Logic Redone
When a user clicks on a text box they are presented with an assistive
example.
When the click off the form js has a blur call to validate data that
passes in the id of the textbox as well as the span above it. If the
text fails to validate then the user is presented with an error with a
solution (in red) inside the span tag. Will be adding a red or green
mark based on pass or fail of the validation.