Showing Data - From From and Radio Button
After all values are totaled, the users image is now shown, as well as
all the answers that got the user to that decision, their user profile
info, and a link to the relative path to copy the said image.
Icons now appear next to the text boxes if the user passes or fails a
text box field.